Split `AppTest` into `StructureTest` and `ElementOriginTest`
